add header for Inode, move *.hpp to includes
