+++ src/includes/InodeMap.hpp
@@ -2,12 +2,17 @@
 
 #include <iostream>
 #include <unordered_map>
-#include "Inode.cpp"
+#include "Inode.hpp"
 
 
 typedef std::unordered_map<size_t, Inode> inode_hashmap_t;
 typedef std::string str_t;
 
+
+/**
+* @brief Class for containing file inode instances.
+* @brief Implemented via unordered_map (aka hash map).
+*/
 class InodeMap {
 
 public:
